split partitioning code into seperate file
